django_library/requirements-dev.txt
mixer==7.2.2
factory-boy==3.3.3